pyproject.toml
[project]
name = "livrable-p8"
version = "0.1.0"
description = ""
authors = [
    {name = "S-S-Zheng",email = "zheng.shipo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy (<2.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "imbalanced-learn (>=0.14.1,<0.15.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "fastapi (>=0.115.0)",
    "uvicorn[standard] (>=0.23.0)",
    "pydantic (>=2.6.0)",
    "pydantic-settings (>=2.6.1)",
    "python-multipart (>=0.0.9)",
    "PyYAML (>=6.0)",
    "python-dotenv (>=1.1.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "sqlalchemy (>=2.0)",
    "psycopg2-binary (>=2.9)",
    "scikit-learn (>=1.8.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "livrable_p8", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
# Notebooks & Analyse
notebook = ">=7.5.3,<8.0.0"
ipykernel = ">=7.2.0,<8.0.0"
matplotlib = ">=3.10.8,<4.0.0"
ipywidgets = ">=8.1.8,<9.0.0"
# Qualité de code & Tests
ruff = ">=0.15.2,<0.16.0"
httpx = ">=0.27.0"
pytest = "==8.3.4"
pytest-cov = "==5.0.0"
pytest-dotenv = "==0.5.2"
pytest-asyncio = ">=1.3.0,<2.0.0"
# Documentation
sphinx = ">=7.0"
furo = ">=2023.9.10"
myst-parser = ">=2.0.0"

[tool.ruff]
line-length = 100
target-version = "py312" # ou votre version de python

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B"]
ignore = [
    "E203",
    "B008", # Autorise les appels de fonctions dans les arguments (FastAPI Depends)
    # "E501", # Optionnel : Ignore les lignes trop longues si tu préfères les gérer manuellement
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B017"] # Autorise pytest.raises(Exception) dans les tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"